--- composition/index.ts
@@ -1,4 +1,0 @@
-import useRequest from "./src/useRequest";
-
-// eslint-disable-next-line import/prefer-default-export
-export { useRequest };
