+++ src/index.ts
@@ -1,0 +1,3 @@
+import useRequest from "./useRequest";
+
+export default { useRequest };
